+++ CodeWars/Algorithms/MultiplesOf3And5/Java/src/SolutionTest.java
@@ -1,0 +1,14 @@
+/**
+ * Created by cajetan on 19/07/2017.
+ */
+
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+
+public class SolutionTest {
+    @Test
+    public void test() {
+        assertEquals(23, new Solution().solution(10));
+    }
+
+}